backend/requirements.txt
aiofile==3.9.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
attrs==25.4.0
Authlib==1.6.8
azure-core==1.38.2
azure-identity==1.25.2
beartype==0.22.9
cachetools==7.0.1
caio==0.9.25
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cryptography==46.0.3
cyclopts==4.6.0
distro==1.9.0
dnspython==2.8.0
docstring_parser==0.17.0
docutils==0.22.4
ecdsa==0.19.1
email-validator==2.3.0
exceptiongroup==1.3.1
fastapi==0.124.0
fastmcp==3.0.2
frozenlist==1.8.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
hyperframe==6.1.0
idna==3.11
importlib_metadata==8.7.1
jaraco.classes==3.4.0
jaraco.context==6.1.0
jaraco.functools==4.4.0
jiter==0.12.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.25.1
jsonschema-path==0.4.2
jsonschema-specifications==2025.9.1
keyring==25.7.0
langchain==1.1.3
langchain-core==1.1.3
langchain-mcp-adapters==0.2.1
langchain-openai==1.1.1
langgraph==1.0.4
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.5
langgraph-sdk==0.2.14
langsmith==0.4.56
markdown-it-py==4.0.0
mcp==1.26.0
mdurl==0.1.2
microsoft-kiota-abstractions==1.9.8
microsoft-kiota-authentication-azure==1.9.8
microsoft-kiota-http==1.9.8
microsoft-kiota-serialization-form==1.9.8
microsoft-kiota-serialization-json==1.9.8
microsoft-kiota-serialization-multipart==1.9.8
microsoft-kiota-serialization-text==1.9.8
more-itertools==10.8.0
msal==1.35.0
msal-extensions==1.3.1
msgraph-core==1.3.8
msgraph-sdk==1.55.0
multidict==6.7.1
openai==2.9.0
openapi-pydantic==0.5.1
opentelemetry-api==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
orjson==3.11.5
ormsgpack==1.12.0
packaging==25.0
pathable==0.5.0
platformdirs==4.9.2
propcache==0.4.1
py-key-value-aio==0.4.4
pyasn1==0.6.1
pyclean==3.4.0
pycparser==2.23
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.10.1
pyperclip==1.11.0
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.20
pywin32==311
pywin32-ctypes==0.2.3
PyYAML==6.0.3
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.3.3
rich-rst==1.3.2
rpds-py==0.30.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
sse-starlette==3.0.3
starlette==0.50.0
std-uritemplate==2.0.8
tenacity==9.1.2
tiktoken==0.12.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.1
uuid_utils==0.12.0
uvicorn==0.38.0
watchfiles==1.1.1
websockets==15.0.1
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
zstandard==0.25.0